changed FeedTopChoice, to -- FeedTopBar
