+++ frontend/components/modules/Blogg/FeedTopBarFeedOptionButton.test.ts
@@ -1,11 +1,11 @@
-import FeedTopChoiceFollowingOptionButton from './FeedTopChoiceFollowingOptionButton.vue';          
+import FeedTopBarFeedOptionButton from './FeedTopBarFeedOptionButton.vue';
 import { shallowMount } from '@vue/test-utils';
 import { describe, expect, test, beforeEach, afterEach } from 'vitest';
 
 let wrapper: any;
 
 const factory = () => {
-    return shallowMount(FeedTopChoiceFollowingOptionButton, {
+    return shallowMount(FeedTopBarFeedOptionButton, {
         global: {
             plugins: [],
             components: {},
@@ -38,7 +38,7 @@ describe('Testing the all posts feed button', () => {
     })
     test('Should have basebutton', () => {
         wrapper = factory()
-        expect(wrapper.find("[data-test='following_option']").exists()).toBe(true)
+        expect(wrapper.find("[data-test='feed_option']").exists()).toBe(true)
         
     })
     test('Should have a selected prop', () => {
@@ -47,15 +47,15 @@ describe('Testing the all posts feed button', () => {
     })
     test('The button should emit when clicked', async () => {
         wrapper = factory()
-        const button = wrapper.find("[data-test='following_option']")
+        const button = wrapper.find("[data-test='feed_option']")
         
         await button.trigger("click")
-        expect(wrapper.emitted("showFollowingPosts")).toBeTruthy()
+        expect(wrapper.emitted("showAllFeedPosts")).toBeTruthy()
     })
     test('The button should have the correct text', async () => {
         wrapper = factory()
-        const button = wrapper.find("[data-test='following_option']")
-        expect(button.attributes("text")).toBe("Following")
+        const button = wrapper.find("[data-test='feed_option']")
+        expect(button.attributes("text")).toBe("Feed")
     })
     test('Should match snapshot', () => {
         wrapper = factory()
